fixed bugs with percussive strike durations
The specific bug was using us instead of beats in the
duration argument to the s method, which resulted in
a hugely long duration for strikes created with the p
methods.  Tried restoring the original behavior, where
percussive strikes were 5ms, and this does not work with
some soundfonts (including the TR-808 soundfont.)  So,
now the default duration for percussive strikes is .25
beats, and it can be changed by the (new) pdur composer
method.

Hacked around with Demo4, have a rhythm pattern going.